+++ features/step_definitions/connect.rb
@@ -8,3 +8,10 @@ When /^the( \w+)? client connects(?: with (\w+) TLS)?$/ do
   client(client_name).start
   client(client_name).connect(server.host, server.port)
 end
+
+When /^the (\d+)rd client tries to connect$/ do |client_name|
+  client(client_name).start
+  capture_error do
+    client(client_name).connect(server.host, server.port)
+  end
+end