+++ features/support/test_server.rb
@@ -222,6 +222,7 @@ class TestServer
   def_delegator :@server, :'server_name='
   def_delegator :@server, :'session_timeout='
   def_delegator :@server, :'tls='
+  def_delegator :@server, :'max_connections='
 
   def_delegator :@driver, :'append='
   def_delegator :@driver, :'delete='